requirements.txt
Django<1.9
neo4jdb==0.0.8
